Introduce defgenerator
This is a very simple macro, without much of anything with regards to
"non-conforming" forms.  This means that either it works as a proxy for a
(format stream "...") function, or you have to manually define the defmethod
yourself.  I'll add more possibilities for this macro later, but now I think we
gain a little in readability, since there are a million types to cover.