Huge rework
I have co-located all the parts of the ast.  Now the node, parser and generator
is inside its own macro (defgql), mostly for convenience.  The benefit is
twofold:

1. No need to jump between files or grep to see now a node is defined.
2. The macro signals an error if there is no node, parser or generator present

I did not intern any symbols or do any magic at all inside this macro, because
I'm not married to the api yet, and I want it to be a little verbose until I get
closer.  However, the colocating should help when expanding that macro down the
line.