Add get-operation
The first steps of execute-request are being taken.  It is too boring just to
implement the spec from top to bottom, so I'll try to implement as much as I
need to make a functional client-server request.  Then fill in more of the
blanks as I go.  Maybe I'll get some help by then...