First iteration of input/output-type-p
Don't feel confident at all about this.  It looks like it should be able to
resolve an unwrapped type here, and I don't think they talk about the
definitions here, but whether the current value of the type is in one of those
contexts.

For example if type is `[Dog]` where `Dog` is a variant on a
`union-type-definition`, then we return true?  So this code does only reflect
the general algorithm, right?