Flesh out complete-value a little more
This looks much more like it.  Now we recur on the wrapper type, and check what
the actual schema returns for us.  We need to know what type of definitions we
are, and we are getting closer.  I'm also liking this pattern a bit more.
Though I think we can get more out of the type system of common lisp down the
line.  Getting fun!