Megacheckpoint!
Ok, so we actually can run queries now!

This is really, really great news! This means that the structure of the
algorithms work, and that we actually can return values, given a perfect world.
Still lots of validation missing, and many, many bugs for the execution, but
this is a GREAT start.  We don't yet resolve values, either.

But yay!