Introduce resolve
I'm not completely sure how to deal with this, but the spec mentions getting the
correct resolver somehow.  This is something that could be handled by the
generic function infrastucture just as well now, isn't it?  I'm thinking this:

The spec describes a hash table of functions, which means we specialize on few
things.  This should just as well be covered by the defgenerics, since we can
specialize on generic stuff.  Let's just see where this takes us?