Introduce bool and actually 'get' the resolve
Ok, so what's changed is that we now need to be able to distinguish between nil,
null, t, true and false.  To do this, I have introduced `'true` and `'false`,
for use in the resolvers.  It is crucial that we actually get to map things over
to this type, so that we can handle value coercing properly.  This still needs
documenting, but for now I need to use it more.

In addition, I finally understood what the object-value is for.  So when we get
the value from the database we typically map it over to something, like a clos
object or struct.  Now the properties on this object can be queried by the field
resolvers.  This means that we overfetch when we ask for an object with 10
fields, but only want 2 of them in return.  I think this is intended behavior.

Does this mean we need to get _all_ the data from the root, even for things that
are deeply nested? I think so.